feat(host): install TanStack Router + scaffold route tree (#147 TC3)
Adds @tanstack/react-router + @tanstack/router-devtools. Wires code-based routing (not file-based via the TanStack Vite plugin — Bakin builds through Bun.build() which has no equivalent plugin).

Layout:
- packages/host/src/routes/__root.tsx — root route component: Providers + Header + LayoutShell wrapping <Outlet />. Theme-overrides injection from bakin.config.theme moved here from the removed App.tsx. Wraps everything.
- packages/host/src/routes/index.tsx — placeholder home route ('/'). TC4 ports the real content.
- packages/host/src/router.ts — createRouter() with the route tree. Declare module augments @tanstack/react-router's Register interface for type-safe navigation.
- packages/host/src/main.tsx — swapped <App/> for <RouterProvider router={router}/>.
- packages/host/src/App.tsx — removed. Superseded by __root.tsx; main.tsx drives the router directly.

TC4-TC24 add 21 more leaf routes under __root via addChildren.

Verification:
- bunx tsc --noEmit clean
- bun run build:host — succeeds
- bunx vitest run — 2983 passed, 1 skipped (unchanged)
