+++ package.json
@@ -40,6 +40,8 @@
     "@dnd-kit/react": "^0.3.2",
     "@hookform/resolvers": "^5.2.2",
     "@modelcontextprotocol/sdk": "^1.27.0",
+    "@tanstack/react-router": "^1.168.23",
+    "@tanstack/router-devtools": "^1.166.13",
     "@types/js-yaml": "^4.0.9",
     "@types/nodemailer": "^7.0.11",
     "@xyflow/react": "^12.10.1",
